# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/support/threads": {
      "post": {
        "description": "Crée le thread ET envoie le premier message en une seule requête.",
        "operationId": "SupportController_createThread",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThreadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Discussion créée avec le premier message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ouvrir une nouvelle discussion de support",
        "tags": [
          "Support"
        ]
      },
      "get": {
        "operationId": "SupportController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister mes discussions de support",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/support/threads/{threadId}": {
      "get": {
        "operationId": "SupportController_findOne",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d'une discussion (avec tous les messages)",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/support/admin/threads": {
      "get": {
        "operationId": "SupportController_findAllAdmin",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "RESOLVED",
                "CLOSED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Lister toutes les discussions avec filtre par statut",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/support/admin/threads/{threadId}": {
      "get": {
        "operationId": "SupportController_findOneAdmin",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Détail complet d'une discussion",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/support/admin/threads/{threadId}/status": {
      "patch": {
        "description": "\n                Cycle autorisé :\n                OPEN → IN_PROGRESS → RESOLVED → CLOSED\n                RESOLVED ou CLOSED → OPEN (réouverture)\n\n                Si une note est fournie, un message système est automatiquement ajouté au thread.\n    ",
        "operationId": "SupportController_updateStatus",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "query",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThreadStatusDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Transition de statut invalide"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Changer le statut d'une discussion",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/support/threads/{threadId}/messages": {
      "get": {
        "operationId": "SupportController_getMessages",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les messages d'une discussion (paginé)",
        "tags": [
          "Support"
        ]
      },
      "post": {
        "description": "\n            - Utilisateur → senderType: USER (par défaut)\n            - Agent support → senderType: AGENT\n            - Bot → senderType: BOT (senderId = null)\n\n            Si un agent répond à un thread OPEN, le statut passe automatiquement à IN_PROGRESS.\n            Impossible d'envoyer dans un thread CLOSED (sauf admin).\n    ",
        "operationId": "SupportController_sendMessage",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Envoyer un message dans une discussion",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/support/admin/threads/{threadId}/messages": {
      "post": {
        "operationId": "SupportController_sendMessageAdmin",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Envoyer un message en tant qu'agent ou bot",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/support/threads/{threadId}/messages/{messageId}": {
      "delete": {
        "operationId": "SupportController_deleteMessage",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer son propre message",
        "tags": [
          "Support"
        ]
      }
    }
  },
  "info": {
    "title": "projet-x-service",
    "description": "API du projet-x-service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateThreadDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Sujet de la discussion",
            "example": "Impossible d'exporter mon livre en PDF"
          },
          "message": {
            "type": "string",
            "description": "Premier message décrivant le problème",
            "example": "Bonjour, quand je clique sur Exporter en PDF, je reçois une erreur 500..."
          }
        },
        "required": [
          "subject",
          "message"
        ]
      },
      "UpdateThreadStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Nouveau statut du thread",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED",
              "CLOSED"
            ],
            "example": "IN_PROGRESS"
          },
          "note": {
            "type": "string",
            "description": "Note interne sur le changement de statut",
            "example": "Pris en charge par l'équipe technique"
          }
        },
        "required": [
          "status"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Contenu du message",
            "example": "Merci pour votre retour, nous allons investiguer."
          },
          "senderType": {
            "type": "string",
            "description": "Type d'expéditeur (USER par défaut)",
            "enum": [
              "USER",
              "AGENT",
              "BOT"
            ],
            "default": "USER"
          },
          "senderId": {
            "type": "string",
            "description": "ID de l'expéditeur (null pour BOT)",
            "example": "clxAGENT001"
          }
        },
        "required": [
          "content"
        ]
      }
    }
  }
}